Csudma: Value is not of appropriate type.
This patch fix below MISRA-C violation in xcsudma.c
93 S
Value is not of appropriate type.
MISRA-C:2012/AMD1/TC1 R.10.1,R.10.3,R.10.4,R.10.5,R.11.1


Acked-for-series: Appana Durga Kedareswara rao <appana.durga.rao@xilinx.com>

+++ XilinxProcessorIPLib/drivers/csudma/src/xcsudma.c
@@ -100,7 +100,7 @@ s32 XCsuDma_CfgInitialize(XCsuDma *InstancePtr, XCsuDma_Config *CfgPtr,
 						sizeof(XCsuDma_Config));
 	InstancePtr->Config.BaseAddress = EffectiveAddr;
 
-	if (InstancePtr->Config.DmaType == XCSUDMA_DMATYPEIS_CSUDMA) {
+	if (InstancePtr->Config.DmaType == (u8)XCSUDMA_DMATYPEIS_CSUDMA) {
 		/* Reset CSUDMA */
 		XCsuDma_Reset();
 	}
